requirements.txt
psycopg2==2.9.9
pydrive==1.3.1
python-dotenv==1.0.1
python-telegram-bot==21.6
schedule==1.2.2
